requirements.txt
torch==1.13.1
torchvision==0.14.1
colorlog
compressai
marshmallow
Pillow
PyYAML
rich
setuptools
tensorboardX
thop
wandb